// packages/astro/dist/core/app/types.d.ts
export type StylesheetAsset =
	| {
			type: 'inline';
			content: string;
	  }
	| {
			type: 'external';
			src: string;
	  };
type ScriptAsset =
	| {
			children: string;
			stage: string;
	  }
	| {
			type: 'inline' | 'external';
			value: string;
	  };
export interface RouteInfo {
	routeData: RouteData;
	file: string;
	links: string[];
	scripts: ScriptAsset[];
	styles: StylesheetAsset[];
}